Changed type of gridsize to size_t.

+++ src/Collgrid.cc
@@ -204,7 +204,7 @@ int genGrid(int ngrids, int nfiles, ens_file_t *ef, bool ginit, int igrid, int n
   if ( nx*ny != nfiles ) cdoAbort("Number of input files (%d) and number of blocks (%dx%d) differ!", nfiles, nx, ny);
  
   int xsize2 = 0;
-  for ( size_t i = 0; i < nx; ++i ) xsize2 += xsize[xyinfo[i].id];
+  for ( int i = 0; i < nx; ++i ) xsize2 += xsize[xyinfo[i].id];
   int ysize2 = 0;
   for ( int j = 0; j < ny; ++j ) ysize2 += ysize[xyinfo[j*nx].id];
   if ( cdoVerbose ) cdoPrint("xsize2 %d  ysize2 %d", xsize2, ysize2);
@@ -224,7 +224,7 @@ int genGrid(int ngrids, int nfiles, ens_file_t *ef, bool ginit, int igrid, int n
   int *yoff = (int*) Malloc((ny+1)*sizeof(int));
 
   xoff[0] = 0;
-  for ( size_t i = 0; i < nx; ++i )
+  for ( int i = 0; i < nx; ++i )
     {
       int idx = xyinfo[i].id;
       if ( lregular ) memcpy(xvals2+xoff[i], xvals[idx], xsize[idx]*sizeof(double));
