+++ src/Invert.cc
@@ -50,8 +50,8 @@ void invertLonDes(int vlistID)
       if ( gridInqXvals(gridID1, NULL) )
 	{
 	  size_t nlon = gridInqXsize(gridID1);
-	  int nlat = gridInqYsize(gridID1);
-	  int size = (gridtype == GRID_CURVILINEAR) ? nlon*nlat : nlon;
+	  size_t nlat = gridInqYsize(gridID1);
+	  size_t size = (gridtype == GRID_CURVILINEAR) ? nlon*nlat : nlon;
 
 	  double *xv1 = (double*) Malloc(size*sizeof(double));
 	  double *xv2 = (double*) Malloc(size*sizeof(double));
@@ -60,13 +60,13 @@ void invertLonDes(int vlistID)
 
 	  if ( gridtype == GRID_CURVILINEAR )
 	    {
-	      for ( int ilat = 0; ilat < nlat; ilat++ )
-		for ( int ilon = 0; ilon < nlon; ilon++ )
+	      for ( size_t ilat = 0; ilat < nlat; ilat++ )
+		for ( size_t ilon = 0; ilon < nlon; ilon++ )
 		  xv2[ilat*nlon + nlon-ilon-1] = xv1[ilat*nlon + ilon];
 	    }
 	  else
 	    {
-	      for ( int ilon = 0; ilon < nlon; ilon++ )
+	      for ( size_t ilon = 0; ilon < nlon; ilon++ )
 		xv2[nlon-ilon-1] = xv1[ilon];
 	    }
 
@@ -79,9 +79,9 @@ void invertLonDes(int vlistID)
       if ( gridInqXbounds(gridID1, NULL) )
 	{
 	  size_t nlon = gridInqXsize(gridID1);
-	  int nlat = gridInqYsize(gridID1);
+	  size_t nlat = gridInqYsize(gridID1);
 	  int nv   = gridInqNvertex(gridID1);
-	  int size = (gridtype == GRID_CURVILINEAR) ? nv*nlon*nlat : nv*nlon;
+	  size_t size = (gridtype == GRID_CURVILINEAR) ? nv*nlon*nlat : nv*nlon;
 
 	  double *xb1 = (double*) Malloc(size*sizeof(double));
 	  double *xb2 = (double*) Malloc(size*sizeof(double));
@@ -90,14 +90,14 @@ void invertLonDes(int vlistID)
 
 	  if ( gridtype == GRID_CURVILINEAR )
 	    {
-	      for ( int ilat = 0; ilat < nlat; ilat++ )
-		for ( int ilon = 0; ilon < nlon; ilon++ )
+	      for ( size_t ilat = 0; ilat < nlat; ilat++ )
+		for ( size_t ilon = 0; ilon < nlon; ilon++ )
 		  for ( int iv = 0; iv < nv; iv++ )
 		    xb2[ilat*nlon*nv + (nlon-ilon-1)*nv + iv] = xb1[ilat*nlon*nv + ilon*nv + iv];
 	    }
 	  else
 	    {
-		for ( int ilon = 0; ilon < nlon; ilon++ )
+		for ( size_t ilon = 0; ilon < nlon; ilon++ )
 		  {
 		    xb2[nlon*2-ilon*2-1] = xb1[ilon*2];
 		    xb2[nlon*2-ilon*2-2] = xb1[ilon*2+1];
@@ -122,8 +122,8 @@ void invertLatCoord(int gridID)
   if ( gridInqYvals(gridID, NULL) )
     {
       size_t nlon = gridInqXsize(gridID);
-      int nlat = gridInqYsize(gridID);
-      int size = (gridtype == GRID_CURVILINEAR) ? nlon*nlat : nlat;
+      size_t nlat = gridInqYsize(gridID);
+      size_t size = (gridtype == GRID_CURVILINEAR) ? nlon*nlat : nlat;
 
       double *yv1 = (double*) Malloc(size*sizeof(double));
       double *yv2 = (double*) Malloc(size*sizeof(double));
@@ -132,16 +132,16 @@ void invertLatCoord(int gridID)
         {
           gridInqXvals(gridID, yv1);
           
-          for ( int ilat = 0; ilat < nlat; ilat++ )
-            for ( int ilon = 0; ilon < nlon; ilon++ )
+          for ( size_t ilat = 0; ilat < nlat; ilat++ )
+            for ( size_t ilon = 0; ilon < nlon; ilon++ )
               yv2[(nlat-ilat-1)*nlon + ilon] = yv1[ilat*nlon + ilon];
 
           gridDefXvals(gridID, yv2);
           
           gridInqYvals(gridID, yv1);
 
-          for ( int ilat = 0; ilat < nlat; ilat++ )
-            for ( int ilon = 0; ilon < nlon; ilon++ )
+          for ( size_t ilat = 0; ilat < nlat; ilat++ )
+            for ( size_t ilon = 0; ilon < nlon; ilon++ )
               yv2[(nlat-ilat-1)*nlon + ilon] = yv1[ilat*nlon + ilon];
 
           gridDefYvals(gridID, yv2);
@@ -150,7 +150,7 @@ void invertLatCoord(int gridID)
         {
           gridInqYvals(gridID, yv1);
 
-          for ( int ilat = 0; ilat < nlat; ilat++ )
+          for ( size_t ilat = 0; ilat < nlat; ilat++ )
             yv2[nlat-ilat-1] = yv1[ilat];
 
           gridDefYvals(gridID, yv2);
@@ -163,9 +163,9 @@ void invertLatCoord(int gridID)
   if ( gridInqYbounds(gridID, NULL) )
     {
       size_t nlon = gridInqXsize(gridID);
-      int nlat = gridInqYsize(gridID);
+      size_t nlat = gridInqYsize(gridID);
       int nv   = gridInqNvertex(gridID);
-      int size = (gridtype == GRID_CURVILINEAR) ? nv*nlon*nlat : nv*nlat;
+      size_t size = (gridtype == GRID_CURVILINEAR) ? nv*nlon*nlat : nv*nlat;
 
       double *yb1 = (double*) Malloc(size*sizeof(double));
       double *yb2 = (double*) Malloc(size*sizeof(double));
@@ -174,14 +174,14 @@ void invertLatCoord(int gridID)
 
       if ( gridtype == GRID_CURVILINEAR )
         {
-          for ( int ilat = 0; ilat < nlat; ilat++ )
-            for ( int ilon = 0; ilon < nlon; ilon++ )
+          for ( size_t ilat = 0; ilat < nlat; ilat++ )
+            for ( size_t ilon = 0; ilon < nlon; ilon++ )
               for ( int iv = 0; iv < nv; iv++ )
                 yb2[(nlat-ilat-1)*nlon*nv + ilon*nv + iv] = yb1[ilat*nlon*nv + ilon*nv + iv];
         }
       else
         {
-          for ( int ilat = 0; ilat < nlat; ilat++ )
+          for ( size_t ilat = 0; ilat < nlat; ilat++ )
             {
               yb2[nlat*2-ilat*2-1] = yb1[ilat*2];
               yb2[nlat*2-ilat*2-2] = yb1[ilat*2+1];
@@ -223,21 +223,21 @@ static
 void invertLonData(double *array1, double *array2, int gridID1)
 {
   size_t nlon = gridInqXsize(gridID1);
-  int nlat = gridInqYsize(gridID1);
+  size_t nlat = gridInqYsize(gridID1);
 
   if ( nlat > 0 )
     {
       double **field1 = (double **) Malloc(nlat*sizeof(double *));
       double **field2 = (double **) Malloc(nlat*sizeof(double *));
   
-      for ( int ilat = 0; ilat < nlat; ilat++ )
+      for ( size_t ilat = 0; ilat < nlat; ilat++ )
 	{
 	  field1[ilat] = array1 + ilat*nlon;
 	  field2[ilat] = array2 + ilat*nlon;
 	}
 
-      for ( int ilat = 0; ilat < nlat; ilat++ )
-	for ( int ilon = 0; ilon < nlon; ilon++ )
+      for ( size_t ilat = 0; ilat < nlat; ilat++ )
+	for ( size_t ilon = 0; ilon < nlon; ilon++ )
 	  field2[ilat][nlon-ilon-1] = field1[ilat][ilon];
   
       if ( field1 ) Free(field1);
@@ -253,20 +253,20 @@ static
 void invertLatData(double *array1, double *array2, int gridID1)
 {
   size_t nlon = gridInqXsize(gridID1);
-  int nlat = gridInqYsize(gridID1);
+  size_t nlat = gridInqYsize(gridID1);
 
   if ( nlat > 0 )
     {
       double **field1 = (double **) Malloc(nlat*sizeof(double *));
       double **field2 = (double **) Malloc(nlat*sizeof(double *));
   
-      for ( int ilat = 0; ilat < nlat; ilat++ )
+      for ( size_t ilat = 0; ilat < nlat; ilat++ )
 	{
 	  field1[ilat] = array1 + ilat*nlon;
 	  field2[ilat] = array2 + ilat*nlon;
 	}
 
-      for ( int ilat = 0; ilat < nlat; ilat++ )
+      for ( size_t ilat = 0; ilat < nlat; ilat++ )
 	memcpy(field2[nlat-ilat-1], field1[ilat], nlon*sizeof(double));
       
       if ( field1 ) Free(field1);
