+++ src/grid.cc
@@ -722,7 +722,7 @@ void field2regular(int gridID1, int gridID2, double missval, double *array, size
 
 int gridToRegular(int gridID1)
 {
-  int nx = 0;
+  size_t nx = 0;
   double *xvals = NULL;
 
   int gridtype = gridInqType(gridID1);
@@ -1459,7 +1459,7 @@ int gridCurvilinearToRegular(int gridID1)
   if ( gridtype != GRID_CURVILINEAR ) return gridID2;
 
   size_t nx = gridInqXsize(gridID1);
-  int ny = gridInqYsize(gridID1);
+  size_t ny = gridInqYsize(gridID1);
 	
   double *xvals2D = (double*) Malloc(gridsize*sizeof(double));
   double *yvals2D = (double*) Malloc(gridsize*sizeof(double));
@@ -1471,9 +1471,9 @@ int gridCurvilinearToRegular(int gridID1)
   double *yvals = (double*) Malloc(ny*sizeof(double));
 
   for ( size_t i = 0; i < nx; i++ ) xvals[i] = xvals2D[i];
-  for ( int j = 0; j < ny; j++ ) yvals[j] = yvals2D[j*nx];
+  for ( size_t j = 0; j < ny; j++ ) yvals[j] = yvals2D[j*nx];
 
-  for ( int j = 1; j < ny; j++ )
+  for ( size_t j = 1; j < ny; j++ )
     for ( size_t i = 0; i < nx; i++ )
       {
 	if ( fabs(xvals[i] - xvals2D[j*nx+i]) > 1.e-6 )
@@ -1484,8 +1484,8 @@ int gridCurvilinearToRegular(int gridID1)
 	  }
       }
 	
-  for ( int i = 1; i < nx; i++ )
-    for ( int j = 0; j < ny; j++ )
+  for ( size_t i = 1; i < nx; i++ )
+    for ( size_t j = 0; j < ny; j++ )
       {
 	if ( fabs(yvals[j] - yvals2D[j*nx+i]) > 1.e-6 )
 	  {
@@ -1614,7 +1614,7 @@ int gridWeightsOld(int gridID, double *weights)
 	      if ( phi2 > (   M_PI/2) ) phi2 =    M_PI/2;
 	      if ( phi2 < (-1*M_PI/2) ) phi2 = -1*M_PI/2;
 	      double sindphi = sin(phi2)-sin(phi1);
-	      for ( int i = 0; i < nlon; i++ )
+	      for ( size_t i = 0; i < nlon; i++ )
 		{
 		  if ( lons[i] >= lons[0]+360 || fabs(lats[j]) > 90 )
 		    weights[datapoint] = 0;
